Initial implementation of portable PDB (sequence points only)
